Refactor background imports and restructure files
- Moved background-related classes to a new directory structure.
- Updated import paths in multiple files to reflect the new locations.
