+++ src/easydiffraction/Profiles/Experiment.py
@@ -9,8 +9,8 @@ from easyscience.Objects.ObjectClasses import Descriptor
 from easyscience.Objects.ObjectClasses import Parameter
 from gemmi import cif
 
-from easydiffraction.elements.Backgrounds.Point import BackgroundPoint
-from easydiffraction.elements.Backgrounds.Point import PointBackground
+from easydiffraction.job.experiment.backgrounds.point import BackgroundPoint
+from easydiffraction.job.experiment.backgrounds.point import PointBackground
 from easydiffraction.io.cif_reader import background_from_cif_block as background_from_cif
 from easydiffraction.io.cif_reader import data_from_cif_block as data_from_cif
 from easydiffraction.io.cif_reader import parameters_from_cif_block as parameters_from_cif
